// frontend/magic-web/src/opensource/pages/superMagic/pages/Workspace/types.ts
export enum TopicMode {
	/** 通用模式 */
	General = "general",
	/** 聊天模式 */
	Chat = "chat",
	/** 数据分析 */
	DataAnalysis = "data_analysis",
	/** PPT */
	PPT = "ppt",
	/** 研报模式 */
	Report = "report",
	/** 录音总结 */
	RecordSummary = "summary",
	/** 空模式 */
	Empty = "",
	/** 设计模式 */
	Design = "design",
	/** 员工主导模式
	 * 不会在员工列表中存在，只用于创建员工时使用
	 */
	AgentManager = "agent-manager",
}

export enum AgentType {
	Official = 1, // 官方/内置智能体
	Custom = 2, // 自定义智能体
	Public = 3, // 公开智能体
}

export interface ModeModelGroupItem {
	group: ModeModelGroup
	models: ModelItem[]
	model_ids: string[]
	image_models?: ModelItem[]
	image_model_ids: string[]
}

export interface ModeModelGroupItemResponse {
	group: ModeModelGroup
	model_ids: string[]
	image_model_ids: string[]
}

export interface ModeItem {
	agent: {
		type: AgentType
		category: "all" | "frequent"
	}
	mode: {
		id: string
		name: string
		identifier: string
		icon: string
		color: string
		icon_url: string
		icon_type: IconType
		sort: number
		placeholder?: string
		playbooks: SceneItem[]
	}
	groups: ModeModelGroupItem[]
}